refactor: 노선 내 역 찾는 로직 Sections 에 이관
